Moved is_empty function to utils
